Adding "scope" subcommand
Setup a scope subcommand for setting/viewing/unsetting a cloudlet and
blueprint scope for other commands.  In the process, cleaned some files
and added a bit of error handling.  This is not plumbed in anywhere yet,
however, because it is actually somewhat more complex to do so than I
expected, so scope is disabled for the time being.