Serious refactoring of object model
Mostly this is a simplification of the object model, removing classes
and methods that are currently only called in one other place, and using
the Python property pattern.  The intent is to make the code easier to
maintain and more clear for new developers, at some cost to
extensibility.  Also, the pattern/lookup functionality is removed for
the time being (further reducing complexity) in favor of a simpler
cloudlet format for version 1.